delete unnecessary folder
