+++ src/textual/app_components/app_compositor.py
