+++ src/textual/app_components/app_renderer.py
