--- src/textual/textual_app/app.py
@@ -1,45 +1,0 @@
-from textual_app.components.base import AppBase
-from textual_app.components.view import AppView
-from textual_app.components.events import AppEvents
-from textual_app.components.widgets import AppWidgets
-
-class App(AppBase, AppView, AppEvents, AppWidgets):
-    def __init__(self):
-        super().__init__()
-
-    def run(self):
-        # Основная логика запуска приложения
-        self.initialize()
-        self.setup()
-        self.start_main_loop()
-
-    def initialize(self):
-        print("Initializing the application")
-        # Инициализация базовых компонентов, представлений, событий и виджетов
-        super().__init__()
-
-    def setup(self):
-        print("Setting up the application")
-        # Настройка начальных параметров, загрузка конфигураций, создание начальных виджетов и т.д.
-        self.add_widget("MainWidget")
-        self.render()
-
-    def start_main_loop(self):
-        print("Starting the main loop")
-        # Основной цикл обработки событий приложения
-        while True:
-            event = self.get_next_event()
-            if event is None:
-                break
-            self.handle_event(event)
-
-    def get_next_event(self):
-        # Метод для получения следующего события (заглушка для примера)
-        import random
-        if random.random() < 0.1:  # Завершаем цикл с вероятностью 10%
-            return None
-        return "sample_event"
-
-if __name__ == "__main__":
-    app = App()
-    app.run()
