--- src/textual/textual_app/components/__init__.py
@@ -1,27 +1,0 @@
-from .base import AppBase
-from .view import AppView
-from .events import AppEvents
-from .widgets import AppWidgets
-from .widget_base import Widget
-from .widget_layout import WidgetLayout
-from .widget_events import WidgetEvent
-from .widget_helpers import helper_function
-from .screen_base import Screen
-from .screen_layout import ScreenLayout
-from .screen_events import ScreenEvent
-from .screen_helpers import helper_function
-
-__all__ = [
-    "AppBase",
-    "AppView",
-    "AppEvents",
-    "AppWidgets"
-    "Widget",
-    "WidgetLayout",
-    "WidgetEvent",
-    "helper_function"
-    "Screen",
-    "ScreenLayout",
-    "ScreenEvent",
-    "helper_function"
-]
