--- src/textual/textual_app/components/base.py
@@ -1,9 +1,0 @@
-class AppBase:
-    def __init__(self):
-
-        self.name = "Base Application"
-        print("AppBase initialized")
-
-    def run(self):
-
-        print(f"{self.name} is running")
