--- src/textual/textual_app/components/events.py
@@ -1,8 +1,0 @@
-class AppEvents:
-    def __init__(self):
-
-        print("AppEvents initialized")
-
-    def handle_event(self, event):
-
-        print(f"Handling event: {event}")
