--- src/textual/textual_app/components/screen_helpers.py
@@ -1,6 +1,0 @@
-from typing import Any
-
-def helper_function(param: Any) -> Any:
-    """A helper function for screens."""
-    pass
-
