--- src/textual/textual_app/components/screen_layout.py
@@ -1,10 +1,0 @@
-from .geometry import Offset, Region, Size
-
-class ScreenLayout:
-    def __init__(self):
-        pass
-
-    def arrange(self, widgets: list[Widget]) -> None:
-        """Arrange the widgets within the layout."""
-        pass
-
