--- src/textual/textual_app/components/view.py
@@ -1,8 +1,0 @@
-class AppView:
-    def __init__(self):
-
-        print("AppView initialized")
-
-    def render(self):
-
-        print("Rendering the view")
