--- src/textual/textual_app/components/widget_layout.py
@@ -1,10 +1,0 @@
-from .geometry import Offset, Region, Size, Spacing
-
-class WidgetLayout:
-    def __init__(self):
-        pass
-
-    def arrange(self, widgets: list[Widget]) -> None:
-        """Arrange the widgets within the layout."""
-        pass
-
