--- src/textual/textual_app/config/__init__.py
@@ -1,3 +1,0 @@
-from .settings import Settings
-
-__all__ = ["Settings"]
